Create main.yml
What is does is it will tell the Github Actions system to go to the steps in main.yml and build the initial build for our project. Once the Actions finish, we can click on 3 dot (...) top right, then select "Create status badge". You need to copy it (makeup) and paste in the README.md file so you will get good document for our check point if the work work for fail.